clk: renesas: r9a09g057: Add XSPI clock/reset
commit fc7dd515374455f07cdd24b8bad3c7952e812bff upstream.

Add clock and reset entries for the XSPI interface on the R9A09G057 SoC.

While at it, rename CLK_PLLCM33_DIV4_PLLCM33 to CLK_PLLCM33_GEAR to align
with the terminology used in the hardware manual.

Signed-off-by: Lad Prabhakar <prabhakar.mahadev-lad.rj@bp.renesas.com>
Reviewed-by: Geert Uytterhoeven <geert+renesas@glider.be>
Link: https://lore.kernel.org/20250627204237.214635-7-prabhakar.mahadev-lad.rj@bp.renesas.com
Signed-off-by: Geert Uytterhoeven <geert+renesas@glider.be>
[OP: Included FIXED_MOD_CONF_XSPI macro from a separate upstream commit
20e31205ae1 ("clk: renesas: r9a09g056: Add XSPI clock/reset") to fix
the build. Only the macro definition was taken; r9a09g056-cpg.c changes
were skipped as the file doesn't exist in 6.12.y-cip]
Signed-off-by: Ovidiu Panait <ovidiu.panait.rb@renesas.com>
Signed-off-by: Pavel Machek <pavel@denx.de>