ANDROID: KVM: Allow retreiving the backing file of PFNMAP mappings
In order to allow taking a reference on the backing file of a mapping
from the KVM/arm page fault path, plumb the kvm_follow_pfn() logic with
the ability to take a reference on the underlying file and return it to
the caller.

BACKPORT: shoehorn the change into hva_to_pfn() since 6.12 doesn't have
the kvm_follow_pfn() logic.

Bug: 444181977
Change-Id: I37952022ede7848b1d700dc930d608c1ced10744
Signed-off-by: Quentin Perret <qperret@google.com>